requirements/base.txt
arrow==0.10.0
binaryornot==0.4.4
chardet==3.0.4
click==6.7
cookiecutter==1.5.1
dj-database-url==0.4.1
Django==1.11.5
django-admin-bootstrapped==2.5.7
django-authtools==1.6.0
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.4
django-filer==1.2.8
django-mptt==0.8.7
django-polymorphic==1.0.2
django-suit==0.2.25
easy-thumbnails==2.4.2
environ==1.0
future==0.16.0
gunicorn==19.6.0
isort==4.2.15
jedi==0.10.2
Jinja2==2.9.6
jinja2-time==0.2.0
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
poyo==0.4.1
psycopg2==2.7.3.1
pycodestyle==2.3.1
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
Unidecode==0.4.21
Werkzeug==0.12.2
whichcraft==0.4.1
whitenoise==3.2
yapf==0.17.0
